Extract FilePreviewModal into dedicated component file
- Move FilePreviewModal from ChatPanel.tsx to modals/FilePreviewModal.tsx
- Remove unused getFilePreviewUrl helper function (dead code)
- Remove unused imports (ImageFullPreview, PdfFullPreview, HighlightedCode)
  from ChatPanel.tsx as they are now only needed in the modal

This improves code organization by:
- Reducing ChatPanel.tsx from 1816 to 1682 lines
- Following the existing modal organization pattern in components/modals/
- Making the FilePreviewModal component reusable if needed elsewhere
